"Do not disturb the boss"
