+++ stake-pool/program/tests/set_staker.rs
@@ -1,0 +1,176 @@
+#![cfg(feature = "test-bpf")]
+
+mod helpers;
+
+use {
+    borsh::{BorshDeserialize, BorshSerialize},
+    helpers::*,
+    solana_program::{
+        hash::Hash,
+        instruction::{AccountMeta, Instruction},
+    },
+    solana_program_test::*,
+    solana_sdk::{
+        instruction::InstructionError, signature::Keypair, signature::Signer,
+        transaction::Transaction, transaction::TransactionError, transport::TransportError,
+    },
+    spl_stake_pool::{error, id, instruction, state},
+};
+
+async fn setup() -> (BanksClient, Keypair, Hash, StakePoolAccounts, Keypair) {
+    let (mut banks_client, payer, recent_blockhash) = program_test().start().await;
+    let stake_pool_accounts = StakePoolAccounts::new();
+    stake_pool_accounts
+        .initialize_stake_pool(&mut banks_client, &payer, &recent_blockhash)
+        .await
+        .unwrap();
+
+    let new_staker = Keypair::new();
+
+    (
+        banks_client,
+        payer,
+        recent_blockhash,
+        stake_pool_accounts,
+        new_staker,
+    )
+}
+
+#[tokio::test]
+async fn success_set_staker_as_manager() {
+    let (mut banks_client, payer, recent_blockhash, stake_pool_accounts, new_staker) =
+        setup().await;
+
+    let mut transaction = Transaction::new_with_payer(
+        &[instruction::set_staker(
+            &id(),
+            &stake_pool_accounts.stake_pool.pubkey(),
+            &stake_pool_accounts.manager.pubkey(),
+            &new_staker.pubkey(),
+        )
+        .unwrap()],
+        Some(&payer.pubkey()),
+    );
+    transaction.sign(&[&payer, &stake_pool_accounts.manager], recent_blockhash);
+    banks_client.process_transaction(transaction).await.unwrap();
+
+    let stake_pool = get_account(&mut banks_client, &stake_pool_accounts.stake_pool.pubkey()).await;
+    let stake_pool = state::StakePool::try_from_slice(&stake_pool.data.as_slice()).unwrap();
+
+    assert_eq!(stake_pool.staker, new_staker.pubkey());
+}
+
+#[tokio::test]
+async fn success_set_staker_as_staker() {
+    let (mut banks_client, payer, recent_blockhash, stake_pool_accounts, new_staker) =
+        setup().await;
+
+    let mut transaction = Transaction::new_with_payer(
+        &[instruction::set_staker(
+            &id(),
+            &stake_pool_accounts.stake_pool.pubkey(),
+            &stake_pool_accounts.staker.pubkey(),
+            &new_staker.pubkey(),
+        )
+        .unwrap()],
+        Some(&payer.pubkey()),
+    );
+    transaction.sign(&[&payer, &stake_pool_accounts.staker], recent_blockhash);
+    banks_client.process_transaction(transaction).await.unwrap();
+
+    let stake_pool = get_account(&mut banks_client, &stake_pool_accounts.stake_pool.pubkey()).await;
+    let stake_pool = state::StakePool::try_from_slice(&stake_pool.data.as_slice()).unwrap();
+
+    assert_eq!(stake_pool.staker, new_staker.pubkey());
+
+    let mut transaction = Transaction::new_with_payer(
+        &[instruction::set_staker(
+            &id(),
+            &stake_pool_accounts.stake_pool.pubkey(),
+            &new_staker.pubkey(),
+            &stake_pool_accounts.staker.pubkey(),
+        )
+        .unwrap()],
+        Some(&payer.pubkey()),
+    );
+    transaction.sign(&[&payer, &new_staker], recent_blockhash);
+    banks_client.process_transaction(transaction).await.unwrap();
+
+    let stake_pool = get_account(&mut banks_client, &stake_pool_accounts.stake_pool.pubkey()).await;
+    let stake_pool = state::StakePool::try_from_slice(&stake_pool.data.as_slice()).unwrap();
+
+    assert_eq!(stake_pool.staker, stake_pool_accounts.staker.pubkey());
+}
+
+#[tokio::test]
+async fn fail_wrong_manager() {
+    let (mut banks_client, payer, recent_blockhash, stake_pool_accounts, new_staker) =
+        setup().await;
+
+    let mut transaction = Transaction::new_with_payer(
+        &[instruction::set_staker(
+            &id(),
+            &stake_pool_accounts.stake_pool.pubkey(),
+            &new_staker.pubkey(),
+            &new_staker.pubkey(),
+        )
+        .unwrap()],
+        Some(&payer.pubkey()),
+    );
+    transaction.sign(&[&payer, &new_staker], recent_blockhash);
+    let transaction_error = banks_client
+        .process_transaction(transaction)
+        .await
+        .err()
+        .unwrap();
+
+    match transaction_error {
+        TransportError::TransactionError(TransactionError::InstructionError(
+            _,
+            InstructionError::Custom(error_index),
+        )) => {
+            let program_error = error::StakePoolError::SignatureMissing as u32;
+            assert_eq!(error_index, program_error);
+        }
+        _ => panic!("Wrong error occurs while malicious try to set manager"),
+    }
+}
+
+#[tokio::test]
+async fn fail_set_staker_without_signature() {
+    let (mut banks_client, payer, recent_blockhash, stake_pool_accounts, new_staker) =
+        setup().await;
+
+    let data = instruction::StakePoolInstruction::SetStaker
+        .try_to_vec()
+        .unwrap();
+    let accounts = vec![
+        AccountMeta::new(stake_pool_accounts.stake_pool.pubkey(), false),
+        AccountMeta::new_readonly(stake_pool_accounts.manager.pubkey(), false),
+        AccountMeta::new_readonly(new_staker.pubkey(), false),
+    ];
+    let instruction = Instruction {
+        program_id: id(),
+        accounts,
+        data,
+    };
+
+    let mut transaction = Transaction::new_with_payer(&[instruction], Some(&payer.pubkey()));
+    transaction.sign(&[&payer], recent_blockhash);
+    let transaction_error = banks_client
+        .process_transaction(transaction)
+        .await
+        .err()
+        .unwrap();
+
+    match transaction_error {
+        TransportError::TransactionError(TransactionError::InstructionError(
+            _,
+            InstructionError::Custom(error_index),
+        )) => {
+            let program_error = error::StakePoolError::SignatureMissing as u32;
+            assert_eq!(error_index, program_error);
+        }
+        _ => panic!("Wrong error occurs while try to set new manager without signature"),
+    }
+}
